Fix extended TS type in computed field example
Not sure if my TypeScript is rusty here but when I tried to use this generic type example in TypeScript 5.3.2, the resulting dynamic type didn't return a union of the supplied type (User) and the extended fields but instead only returned the extended fields (fullName). 

So it returned an intersection instead of a union which seems like the appropriate thing to do here.